moved the 4 vendor/device translate functions to the utils pciids
